refactored TCP Ibis

+++ src/ibis/impl/messagePassing/ArrayOutputStream.java
@@ -101,7 +101,7 @@ final public class ArrayOutputStream
 	out.close();
     }
 
-    public final int bytesWritten() { 
+    public final long bytesWritten() { 
 	return out.getCount();
     }
